python-backend/pyproject.toml
[project]
name = "ai-passage-creator-backend"
version = "0.0.1"
description = "AI 爆款文章创作器 - Python 后端"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.32.0",
    "python-multipart==0.0.12",
    "sqlalchemy==2.0.36",
    "pymysql==1.1.1",
    "cryptography==43.0.3",
    "databases[mysql]==0.9.0",
    "redis==5.2.0",
    "aioredis==2.0.1",
    "pydantic==2.10.2",
    "pydantic-settings==2.6.1",
    "python-dotenv==1.0.1",
    "passlib[bcrypt]==1.7.4",
    "python-dateutil==2.9.0.post0",
    "openai==1.58.1",
    "httpx==0.28.1",
    "cos-python-sdk-v5==1.9.32",
    "google-genai==1.35.0",
    "beautifulsoup4==4.12.3",
    "lxml==5.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []

[tool.hatch.build.targets.wheel]
packages = ["app"]

